python/mcp_server/requirements.txt
fastmcp>=0.1.0
pyzmq>=25.0.0
pyyaml>=6.0
aiohttp>=3.9.0
msgpack>=1.0.0
